# executor: reduce overhead of copystack in idxlookup workers