also refine the worker pool

<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #56649

Problem Summary: copystack in index lookup workers consume a lot of cpu resources.

![image](https://github.com/user-attachments/assets/2c003c3e-fef8-4bee-8113-867a22f6298d)

### What changed and how does it work?

Growing the stack at the beginning of the worker goroutine. Also refine the implementation of the worker pool.

![image](https://github.com/user-attachments/assets/f074aae9-348a-44d5-8573-89c9c1639885)

And here are some benchmark results.

| Workload  | Threads | TPS (this PR) | TPS (baseline) | Diff |
| --------- | ------: | ------------: | -------------: | ---: |
| idxlookup |      32 |       2797.99 |        2708.88 | 3.3% |
| idxlookup |      64 |       3292.84 |        3123.35 | 5.4% |

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```